backend/requirements.txt
tqdm
langchain
langchain-community
langchain-core
langchain-text-splitters
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
pydantic-settings
faiss-cpu
sentence-transformers
numpy
pandas